Created functions to manage the schedule of the student
Parsed the .ics calendar information into an eventData variable first, then added matching logic for each day of the week to display the university classes properly. Added library date-fns to parse schedule data's hours and dates, seperated the events with unique IDs and created seperate objects to store in the data object array for the chart to display. Adding the 'Няма часове' object in between classes, but currently I have no way of seperating if a student is in one of two classes that start in relatively the same time, so to counter that I have added logic to remove the 'Няма часове' object if that is the case.